# Application/Application/Model/Enums/DictionaryImportingStatus.cs
namespace Application.Model.Enums;

/// <summary>
/// Статус импорта справочника
/// </summary>
public enum DictionaryImportingStatus
{
    Success,
    Failed,
    Processing
}